[REFACT] based-range loop for map and size_t type

+++ include/ontoloGenius/core/ontoGraphs/Graphs/ClassGraph.h
@@ -44,7 +44,7 @@ private:
     if(find)
       return;
 
-    for(unsigned int i = 0; i < vect.size(); i++)
+    for(size_t i = 0; i < vect.size(); i++)
       if(disjoint == vect[i]->value_)
       {
         me->setSteady_disjoint(vect[i]);
